+++ changes.md
@@ -7,3 +7,12 @@
   - adaptive weights
 - content aware gating
 
+
+
+#### Notes;
+1. Self-Attention
+2. Add and Norm
+3. Feed Forward
+2. Co-Attention / Cross Attention 
+3. Add and Norm
+4.